Extract converttodm, converttoroom into dm
